Mixi Game
Vanja and Miksi have already finished everything for this year at their university, so they decided to spend their free time playing a game with a binary sequence A1,A2,…,AN (i.e. a sequence containing only integers 0 and 1).

At the beginning of the game, Vanja and Miksi write down the expression ∗A1∗A2∗A3∗…∗AN (note below that each ∗ is a placeholder and does not denote multiplication).

The players take turns alternately, starting with Vanja. The game lasts N turns. In each turn, the current player must replace the first (leftmost) remaining ∗ by the sign + or − (or, equivalently, by the operation of addition or subtraction).

After the last turn, the boys calculate the resulting value V of the expression. If |V|≥K, the winner is Vanja; otherwise, the winner is Miksi.

Please predict the winner of the game if both players play optimally.