Minimum Good Permutation
A permutation of length n is an array of size n consisting of n distinct integers in the range [1, n]. For example, (3, 2, 4, 1) is a permutation of length 4, but (3, 3, 1, 4) and (2, 3, 4, 5) are not, as (3, 3, 1, 4) contains duplicate elements, and (2, 3, 4, 5) contains elements not in range [1,4].

A permutation p of length n is good if and only if for any 1 ≤ i ≤ n, pi ≠ i.

Please find the lexicographically smallest good permutation p.

Definition for "lexicographically smaller":

For two permutations p and q, we say that p is lexicographically smaller than q if and only if there exists a index 1 ≤ l ≤ n such that:

For any 1 ≤ i < l, pi = qi. Note that if l = 1, this constraint means nothing.
and, pl < ql.
For example, (2, 3, 1, 4) < (2, 3, 4, 1) < (3, 4, 1, 2). The lexicographically smallest permutation is, of course, (1, 2, ..., n), though this one is not good.