Little Chef and Numbers
Little chef has just been introduced to the world of numbers! While experimenting with addition and multiplication operations, the little chef came up with the following problem:

Given an array A of non-negative integers, how many pairs of indices i and j exist such that A[i]*A[j] > A[i]+A[j] where i < j .

Now being a learner, little chef isn't able to solve this problem efficiently and hence turns to you for help.