Max Mex
Chef has a sequence of positive integers A1,A2,…,AN. He wants to choose some elements of this sequence (possibly none or all of them) and compute their MEX, i.e. the smallest positive integer which does not occur among the chosen elements. For example, the MEX of [1,2,4] is 3.

Help Chef find the largest number of elements of the sequence A which he can choose such that their MEX is equal to M, or determine that it is impossible.