Name Reduction
In an attempt to reduce the growing population, Archer was asked to come up with a plan. Archer being as intelligent as he is, came up with the following plan:

If N children, with names C1, C2, ..., CN, are born to parents with names A and B, and you consider C to be the concatenation of all the names of the children, i.e. C = C1 + C2 + ... + CN (where + is concatenation operator), then C should be a substring of one of the permutations of A + B.

You are given the task to verify whether the names parents propose to give their children are in fact permissible by Archer's plan or not.