Box of Chocolates
Chef just got a box of chocolates as his birthday gift. The box contains N chocolates in a row (numbered 1 through N), where N is even. For each valid i, the i-th chocolate has a sweetness value Wi.

Chef wants to eat all the chocolates in the first half of the box and leave all chocolates in the second half uneaten. Since he does not like chocolates that are too sweet, he will be unhappy if at least one of the chocolates he eats has the maximum sweetness among all the chocolates in the box.

A right cyclic shift by k chocolates (0≤k<N) consists of moving the last k chocolates in the row to the beginning in the same order and moving each of the remaining N−k chocolates k places to the right. Before eating the first half of the chocolates, Chef wants to perform some right cyclic shift in such a way that he will not be unhappy after eating them. Find the number of ways to do this, i.e. the number of valid integers k such that if Chef performs the right cyclic shift by k chocolates and then eats the first half of the chocolates in the box, he does not become unhappy.