Increasing Decreasing
Chef received a new sequence A1,A2,…,AN. He does not like arbitrarily ordered sequences, so he wants to permute the elements of A in such a way that it would satisfy the following condition: there is an integer p (1≤p≤N) such that the first p elements of the new (permuted) sequence are strictly increasing and the last N−p+1 elements are strictly decreasing.

Help Chef and find a permutation of the given sequence which satisfies this condition or determine that no such permutation exists. If there are multiple solutions, you may find any one.