Misha and Nice Sets
Misha calls a non-empty set of positive integers nice if it has the following properties:

Each element of the set is between l and r (inclusive).
The greatest positive integer which divides each element of the set is exactly g.
Misha likes nice sets, so she wants to find the size of the largest nice set. Help her solve the problem.